requirements.txt
python-docx
docx2pdf
sqlite3
pycups